Quick fix

+++ scripts/elog_submit.sh
@@ -137,7 +137,9 @@ fi
 
 echo "$MAIN_PY -c ${THIS_CONFIGFILE} -t $TASK"
 $MAIN_PY -c ${THIS_CONFIGFILE} -t $TASK
-rm -f ${THIS_CONFIGFILE}
+if [ ${RUN_NUM} != 'None' ]; then
+  rm -f ${THIS_CONFIGFILE}
+fi
 if [ -f ${TMP_EXE} ]; then chmod +x ${TMP_EXE}; . ${TMP_EXE}; rm -f ${TMP_EXE}; fi
 EOF
